refactor: 이전의 acceptanceTest 삭제
