+++ src/test/java/wooteco/subway/auth/AuthAcceptanceTest.java
@@ -5,102 +5,55 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import wooteco.subway.AcceptanceTest;
-import wooteco.subway.auth.dto.TokenResponse;
+import wooteco.subway.auth.dto.TokenRequest;
 
-import java.util.HashMap;
-import java.util.Map;
-
-import static wooteco.subway.member.MemberAcceptanceTest.회원_생성을_요청;
-import static wooteco.subway.member.MemberAcceptanceTest.회원_정보_조회됨;
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class AuthAcceptanceTest extends AcceptanceTest {
-    private static final String EMAIL = "email@email.com";
-    private static final String PASSWORD = "password";
-    private static final Integer AGE = 20;
 
-    @DisplayName("Bearer Auth")
     @Test
-    void myInfoWithBearerAuth() {
-        // given
-        회원_등록되어_있음(EMAIL, PASSWORD, AGE);
-        TokenResponse tokenResponse = 로그인되어_있음(EMAIL, PASSWORD);
-
-        // when
-        ExtractableResponse<Response> response = 내_회원_정보_조회_요청(tokenResponse);
-
-        // then
-        회원_정보_조회됨(response, EMAIL, AGE);
+    @DisplayName("ID가 Email 형식이 아닐 경우의 로그인 요청")
+    public void loginWhenInvalidEmail() {
+        TokenRequest tokenRequest = new TokenRequest("invalid", "password");
+        ExtractableResponse<Response> response = loginRequest(tokenRequest);
+        assertThat(response.statusCode()).isEqualTo(400);
+        assertThat((String) response.jsonPath().get("error")).isEqualTo("INVALID_INPUT");
     }
 
-    @DisplayName("Bearer Auth 로그인 실패")
     @Test
-    void myInfoWithBadBearerAuth() {
-        회원_등록되어_있음(EMAIL, PASSWORD, AGE);
+    @DisplayName("비밀번호가 blank일 경우의 로그인 요청")
+    public void loginWhenInvalidPassword() {
+        TokenRequest tokenRequest = new TokenRequest("kang@kang.com", " ");
+        ExtractableResponse<Response> response = loginRequest(tokenRequest);
+        assertThat(response.statusCode()).isEqualTo(400);
+        assertThat((String) response.jsonPath().get("error")).isEqualTo("INVALID_INPUT");
+    }
 
-        Map<String, String> params = new HashMap<>();
-        params.put("email", EMAIL + "OTHER");
-        params.put("password", PASSWORD);
+    @Test
+    @DisplayName("비밀번호가 틀린 경우의 로그인 요청")
+    public void loginWhenMisMatchCredential() {
+        TokenRequest tokenRequest = new TokenRequest("email@email.com", "wrongPwd");
+        ExtractableResponse<Response> response = loginRequest(tokenRequest);
+        assertThat(response.statusCode()).isEqualTo(400);
+        assertThat((String) response.jsonPath().get("error")).isEqualTo("MISMATCH_ID_PASSWORD");
+    }
 
-        RestAssured
-                .given().log().all()
+    @Test
+    @DisplayName("정상적인 로그인 요청")
+    public void loginWhenValidRequest() {
+        TokenRequest tokenRequest = new TokenRequest("email@email.com", "password");
+        ExtractableResponse<Response> response = loginRequest(tokenRequest);
+        assertThat(response.statusCode()).isEqualTo(200);
+    }
+
+    public static ExtractableResponse<Response> loginRequest(TokenRequest tokenRequest) {
+        return RestAssured.given()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(params)
-                .when().post("/login/token")
-                .then().log().all()
-                .statusCode(HttpStatus.UNAUTHORIZED.value());
-    }
-
-    @DisplayName("Bearer Auth 유효하지 않은 토큰")
-    @Test
-    void myInfoWithWrongBearerAuth() {
-        TokenResponse tokenResponse = new TokenResponse("accesstoken");
-
-        RestAssured
-                .given().log().all()
-                .auth().oauth2(tokenResponse.getAccessToken())
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/members/me")
-                .then().log().all()
-                .statusCode(HttpStatus.UNAUTHORIZED.value());
-    }
-
-    public static ExtractableResponse<Response> 회원_등록되어_있음(String email, String password, Integer age) {
-        return 회원_생성을_요청(email, password, age);
-    }
-
-    public static TokenResponse 로그인되어_있음(String email, String password) {
-        ExtractableResponse<Response> response = 로그인_요청(email, password);
-        return response.as(TokenResponse.class);
-    }
-
-    public static ExtractableResponse<Response> 로그인_요청(String email, String password) {
-        Map<String, String> params = new HashMap<>();
-        params.put("email", email);
-        params.put("password", password);
-
-        return RestAssured.given().log().all().
-                contentType(MediaType.APPLICATION_JSON_VALUE).
-                body(params).
-                when().
-                post("/login/token").
-                then().
-                log().all().
-                statusCode(HttpStatus.OK.value()).
-                extract();
-    }
-
-    public static ExtractableResponse<Response> 내_회원_정보_조회_요청(TokenResponse tokenResponse) {
-        return RestAssured.given().log().all().
-                auth().oauth2(tokenResponse.getAccessToken()).
-                accept(MediaType.APPLICATION_JSON_VALUE).
-                when().
-                get("/members/me").
-                then().
-                log().all().
-                statusCode(HttpStatus.OK.value()).
-                extract();
+                .body(tokenRequest)
+                .when()
+                .post("/api/login/token")
+                .then().extract();
     }
 }
