+++ src/test/java/wooteco/subway/line/LineAcceptanceTest.java
@@ -6,277 +6,351 @@ import io.restassured.response.Response;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import wooteco.subway.AcceptanceTest;
-import wooteco.subway.line.dto.*;
-import wooteco.subway.station.domain.Station;
-import wooteco.subway.station.dto.StationResponse;
-
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-import java.util.stream.Collectors;
+import wooteco.subway.auth.dto.TokenRequest;
+import wooteco.subway.line.dto.LineRequest;
+import wooteco.subway.line.dto.LineResponse;
+import wooteco.subway.line.dto.SectionRequest;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static wooteco.subway.station.StationAcceptanceTest.지하철역_등록되어_있음;
+import static wooteco.subway.auth.AuthAcceptanceTest.loginRequest;
 
-@DisplayName("지하철 노선 관련 기능")
 public class LineAcceptanceTest extends AcceptanceTest {
-    private StationResponse 강남역;
-    private StationResponse downStation;
-    private LineRequest lineRequest1;
-    private LineRequest lineRequest2;
 
+    @Override
     @BeforeEach
     public void setUp() {
         super.setUp();
-
-        // given
-        강남역 = 지하철역_등록되어_있음("강남역");
-        downStation = 지하철역_등록되어_있음("광교역");
-
-        lineRequest1 = new LineRequest("신분당선", "bg-red-600", 강남역.getId(), downStation.getId(), 10);
-        lineRequest2 = new LineRequest("구신분당선", "bg-red-600", 강남역.getId(), downStation.getId(), 15);
-    }
-
-    @DisplayName("지하철 노선을 생성한다.")
-    @Test
-    void createLine() {
-        // when
-        ExtractableResponse<Response> response = 지하철_노선_생성_요청(lineRequest1);
-
-        // then
-        지하철_노선_생성됨(response);
-    }
-
-    @DisplayName("기존에 존재하는 지하철 노선 이름으로 지하철 노선을 생성한다.")
-    @Test
-    void createLineWithDuplicateName() {
-        // given
-        지하철_노선_등록되어_있음(lineRequest1);
-
-        // when
-        ExtractableResponse<Response> response = 지하철_노선_생성_요청(lineRequest1);
-
-        // then
-        지하철_노선_생성_실패됨(response);
-    }
-
-    @DisplayName("지하철 노선 목록을 조회한다.")
-    @Test
-    void getLines() {
-        // given
-        LineResponse lineResponse1 = 지하철_노선_등록되어_있음(lineRequest1);
-        LineResponse lineResponse2 = 지하철_노선_등록되어_있음(lineRequest2);
-        Station station1 = new Station(강남역.getId(), 강남역.getName());
-        Station station2 = new Station(downStation.getId(), downStation.getName());
-        TotalLineResponse totalLineResponse1 = new TotalLineResponse(
-                lineResponse1.getId(), lineResponse1.getName(), lineResponse1.getColor(),
-                station1, station2, 25);
-        TotalLineResponse totalLineResponse2 = new TotalLineResponse(
-                lineResponse2.getId(), lineResponse2.getName(), lineResponse2.getColor(),
-                station1, station2, 25);
-
-        // when
-        ExtractableResponse<Response> response = 지하철_노선_목록_조회_요청();
-
-        // then
-        지하철_노선_목록_응답됨(response);
-        지하철_노선_목록_포함됨(response, Arrays.asList(totalLineResponse1, totalLineResponse2));
-    }
-
-    @DisplayName("지하철 노선을 조회한다.")
-    @Test
-    void getLine() {
-        // given
-        LineResponse lineResponse = 지하철_노선_등록되어_있음(lineRequest1);
-
-        // when
-        ExtractableResponse<Response> response = 지하철_노선_조회_요청(lineResponse);
-
-        // then
-        지하철_노선_응답됨(response, lineResponse);
-    }
-
-    @DisplayName("지하철 노선을 수정한다.")
-    @Test
-    void updateLine() {
-        // given
-        LineResponse lineResponse = 지하철_노선_등록되어_있음(lineRequest1);
-
-        // when
-        ExtractableResponse<Response> response = 지하철_노선_수정_요청(lineResponse, lineRequest2);
-
-        // then
-        지하철_노선_수정됨(response);
-    }
-
-    @DisplayName("지하철 노선을 제거한다.")
-    @Test
-    void deleteLine() {
-        // given
-        LineResponse lineResponse = 지하철_노선_등록되어_있음(lineRequest1);
-
-        // when
-        ExtractableResponse<Response> response = 지하철_노선_제거_요청(lineResponse);
-
-        // then
-        지하철_노선_삭제됨(response);
+        jdbcTemplate.update("insert into STATION (name) values ('가양역')");
+        jdbcTemplate.update("insert into STATION (name) values ('증미역')");
     }
 
     @Test
-    @DisplayName("전체보기 조회 신규 API")
-    public void findAllTotalLineResponse() {
-        //given
-        StationResponse stationResponse1 = 지하철역_등록되어_있음("가양역");
-        StationResponse stationResponse2 = 지하철역_등록되어_있음("당산역");
-        StationResponse stationResponse3 = 지하철역_등록되어_있음("합정역");
-        LineResponse lineResponse1 = 지하철_노선_등록되어_있음("9호선", "황토색", stationResponse1, stationResponse2, 10);
-        LineResponse linseResponse2 = 지하철_노선_등록되어_있음("2호선", "초록색", stationResponse2, stationResponse3, 20);
-        TransferLineResponse transferLineResponse1 = new TransferLineResponse(2L, "2호선", "초록색");
-        TransferLineResponse transferLineResponse2 = new TransferLineResponse(1L, "9호선", "황토색");
-        SectionResponse sectionResponse1 = new SectionResponse(
-                1L, "가양역", 10, Collections.emptyList()
-        );
-        SectionResponse sectionResponse2 = new SectionResponse(
-                1L, "당산역", 0, Collections.singletonList(transferLineResponse1)
-        );
-        SectionResponse sectionResponse3 = new SectionResponse(
-                2L, "당산역", 20, Collections.singletonList(transferLineResponse2)
-        );
-        SectionResponse sectionResponse4 = new SectionResponse(
-                2L, "합정역", 0, Collections.emptyList()
-        );
+    @DisplayName("유효하지 않은 토큰일 경우의 노선 삽입")
+    public void createLineWhenInvalidToken() {
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        LineRequest request = new LineRequest("9호선", "황토색", 1L, 2L, 10);
+        ExtractableResponse<Response> response = createLineRequest(request, "asdf");
 
-        LineMapResponse lineMapResponse1 = new LineMapResponse(
-                1L, "9호선", "황토색", 10, Arrays.asList(sectionResponse1, sectionResponse2)
-                );
-        LineMapResponse lineMapResponse2 = new LineMapResponse(
-                2L, "2호선", "초록색", 20, Arrays.asList(sectionResponse3, sectionResponse4)
-        );
-        List<LineMapResponse> lineMapResponses = Arrays.asList(lineMapResponse1, lineMapResponse2);
-
-        //when
-        ExtractableResponse<Response> response = 지하철_전체_보기_요청();
-
-        //then
-        지하철_전체_보기_응답됨(response);
-        지하철_전체_보기_일치함(response, lineMapResponses);
+        assertThat(response.statusCode()).isEqualTo(401);
+        assertThat((String) response.jsonPath().get("error")).isEqualTo("INVALID_TOKEN");
     }
 
-    public static LineResponse 지하철_노선_등록되어_있음(String name, String color, StationResponse upStation, StationResponse downStation, int distance) {
-        LineRequest lineRequest = new LineRequest(name, color, upStation.getId(), downStation.getId(), distance);
-        return 지하철_노선_등록되어_있음(lineRequest);
+    @Test
+    @DisplayName("유효하지 않은 이름일 경우의 노선 삽입")
+    public void createLineWhenInvalidName() {
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        LineRequest request = new LineRequest("!@#", "황토색", 1L, 2L, 10);
+        ExtractableResponse<Response> response = createLineRequest(request, accessToken);
+
+        assertThat(response.statusCode()).isEqualTo(400);
+        assertThat((String) response.jsonPath().get("error")).isEqualTo("INVALID_NAME");
     }
 
-    public static LineResponse 지하철_노선_등록되어_있음(LineRequest lineRequest) {
-        return 지하철_노선_생성_요청(lineRequest).as(LineResponse.class);
+    @Test
+    @DisplayName("유효하지 않은 색깔일 경우의 노선 삽입")
+    public void createLineWhenInvalidColor() {
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        LineRequest request = new LineRequest("9호선", " ", 1L, 2L, 10);
+        ExtractableResponse<Response> response = createLineRequest(request, accessToken);
+
+        assertThat(response.statusCode()).isEqualTo(400);
+        assertThat((String) response.jsonPath().get("error")).isEqualTo("INVALID_NAME");
     }
 
-    public static ExtractableResponse<Response> 지하철_전체_보기_요청() {
-        return RestAssured
-                .given().log().all()
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/lines/map")
-                .then().log().all()
-                .extract();
+    @Test
+    @DisplayName("유효하지 않은 거리일 경우의 노선 삽입")
+    public void createLineWhenInvalidDistance() {
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        LineRequest request = new LineRequest("9호선", "황토색", 1L, 2L, 0);
+        ExtractableResponse<Response> response = createLineRequest(request, accessToken);
+
+        assertThat(response.statusCode()).isEqualTo(400);
+        assertThat((String) response.jsonPath().get("error")).isEqualTo("INVALID_DISTANCE");
     }
 
-    public static ExtractableResponse<Response> 지하철_노선_생성_요청(LineRequest params) {
-        return RestAssured
-                .given().log().all()
+    @Test
+    @DisplayName("중복되는 노선 이름일 경우의 노선 삽입")
+    public void createLineWhenDuplicatedLineName() {
+        jdbcTemplate.update("insert into LINE (name, color, extraFare) values ('9호선', '황토색', 30)");
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        LineRequest request = new LineRequest("9호선", "황토색", 1L, 2L, 30);
+        ExtractableResponse<Response> response = createLineRequest(request, accessToken);
+
+        assertThat(response.statusCode()).isEqualTo(400);
+        assertThat((String) response.jsonPath().get("error")).isEqualTo("DUPLICATED_LINE_NAME");
+    }
+
+    @Test
+    @DisplayName("존재하지 않는 역이 포함된 경우의 노선 삽입")
+    public void createLineWhenNonExistStation() {
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        LineRequest request = new LineRequest("9호선", "황토색", 10L, 2L, 30);
+        ExtractableResponse<Response> response = createLineRequest(request, accessToken);
+
+        assertThat(response.statusCode()).isEqualTo(400);
+        assertThat((String) response.jsonPath().get("error")).isEqualTo("NO_SUCH_STATION");
+    }
+
+    @Test
+    @DisplayName("상행 하행역이 같은 경우의 노선 삽입")
+    public void createLineWhenSameBothStations() {
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        LineRequest request = new LineRequest("9호선", "황토색", 1L, 1L, 30);
+        ExtractableResponse<Response> response = createLineRequest(request, accessToken);
+
+        assertThat(response.statusCode()).isEqualTo(400);
+        assertThat((String) response.jsonPath().get("error")).isEqualTo("SAME_STATIONS_IN_SAME_SECTION");
+    }
+
+    @Test
+    @DisplayName("정상적인 노선 삽입")
+    public void createLine() {
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        LineRequest request = new LineRequest("9호선", "황토색", 1L, 2L, 30);
+        ExtractableResponse<Response> response = createLineRequest(request, accessToken);
+
+        LineResponse lineResponse = response.body().as(LineResponse.class);
+        assertThat(response.statusCode()).isEqualTo(201);
+        assertThat((String) response.jsonPath().get("name")).isEqualTo("9호선");
+        assertThat((String) response.jsonPath().get("color")).isEqualTo("황토색");
+    }
+
+    @Test
+    @DisplayName("유효하지 않은 토큰일 경우의 모든 노선 목록")
+    public void totalLineResponseWhenInvalidToken() {
+        LineRequest request = new LineRequest("9호선", "황토색", 1L, 2L, 30);
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        createLineRequest(request, accessToken);
+        ExtractableResponse<Response> response = findAllRequest("asdf");
+
+        assertThat(response.statusCode()).isEqualTo(401);
+        assertThat((String) response.jsonPath().get("error")).isEqualTo("INVALID_TOKEN");
+    }
+
+    @Test
+    @DisplayName("정상적인 모든 노선 목록")
+    public void totalLineResponse() {
+        LineRequest request = new LineRequest("9호선", "황토색", 1L, 2L, 30);
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        createLineRequest(request, accessToken);
+        ExtractableResponse<Response> response = findAllRequest(accessToken);
+
+        assertThat(response.statusCode()).isEqualTo(200);
+        assertThat(response.jsonPath().getList("id").get(0)).isEqualTo(1);
+        assertThat(response.jsonPath().getList("name").get(0)).isEqualTo("9호선");
+        assertThat(response.jsonPath().getList("color").get(0)).isEqualTo("황토색");
+        assertThat(response.jsonPath().getList("startStation.id").get(0)).isEqualTo(1);
+        assertThat(response.jsonPath().getList("endStation.id").get(0)).isEqualTo(2);
+        assertThat(response.jsonPath().getList("distance").get(0)).isEqualTo(30);
+    }
+
+    @Test
+    @DisplayName("존재하지 않는 노선에 대한 id를 통한 노선 조회")
+    public void findByIdWhenNotExist() {
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        ExtractableResponse<Response> response = findByIdRequest(1L, accessToken);
+
+        assertThat(response.statusCode()).isEqualTo(400);
+        assertThat((String) response.jsonPath().get("error")).isEqualTo("NO_SUCH_LINE");
+    }
+
+    @Test
+    @DisplayName("정상적인 id를 통한 노선 조회")
+    public void findById() {
+        LineRequest request = new LineRequest("9호선", "황토색", 1L, 2L, 30);
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        createLineRequest(request, accessToken);
+        ExtractableResponse<Response> response = findByIdRequest(1L, accessToken);
+
+        assertThat(response.statusCode()).isEqualTo(200);
+        assertThat((String) response.jsonPath().get("name")).isEqualTo("9호선");
+        assertThat((String) response.jsonPath().get("color")).isEqualTo("황토색");
+    }
+
+    @Test
+    @DisplayName("중복되는 이름으로 노선 수정")
+    public void updateLineWhenDuplicatedName() {
+        LineRequest request = new LineRequest("9호선", "황토색", 1L, 2L, 30);
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        createLineRequest(request, accessToken);
+        LineRequest request2 = new LineRequest("8호선", "초록색", 1L, 2L, 30);
+        createLineRequest(request2, accessToken);
+
+        LineRequest newlyRequest = new LineRequest("9호선", "초록색", 1L, 2L, 30);
+        ExtractableResponse<Response> response = updateLineRequest(2L, newlyRequest, accessToken);
+
+        assertThat(response.statusCode()).isEqualTo(400);
+        assertThat((String) response.jsonPath().get("error")).isEqualTo("DUPLICATED_LINE_NAME");
+    }
+
+    @Test
+    @DisplayName("정상적인 노선 수정")
+    public void updateLine() {
+        LineRequest request = new LineRequest("9호선", "황토색", 1L, 2L, 30);
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        createLineRequest(request, accessToken);
+
+        LineRequest newlyRequest = new LineRequest("9호선", "초록색", 1L, 2L, 30);
+        ExtractableResponse<Response> response = updateLineRequest(1L, newlyRequest, accessToken);
+
+        assertThat(response.statusCode()).isEqualTo(200);
+    }
+
+    @Test
+    @DisplayName("존재하지 않는 노선을 삭제")
+    public void deleteLineWhenNotExist() {
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        ExtractableResponse<Response> response = deleteLineRequest(1L, accessToken);
+
+        assertThat(response.statusCode()).isEqualTo(400);
+        assertThat((String) response.jsonPath().get("error")).isEqualTo("NO_SUCH_LINE");
+    }
+
+    @Test
+    @DisplayName("정상적인 노선 삭제")
+    public void deleteLine() {
+        LineRequest request = new LineRequest("9호선", "황토색", 1L, 2L, 30);
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        createLineRequest(request, accessToken);
+
+        ExtractableResponse<Response> response = deleteLineRequest(1L, accessToken);
+
+        Boolean result = jdbcTemplate.queryForObject("select exists (select * from LINE where name = '9호선')", Boolean.class);
+        assertThat(response.statusCode()).isEqualTo(204);
+        assertThat(result).isFalse();
+    }
+
+    @Test
+    @DisplayName("두 역 다 이미 노선에 존재하는 경우의 구간 추가")
+    public void addSectionWhenBothStationsIncludeInLine() {
+        LineRequest request = new LineRequest("9호선", "황토색", 1L, 2L, 30);
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        createLineRequest(request, accessToken);
+
+        SectionRequest sectionRequest = new SectionRequest(1L, 2L, 10);
+        ExtractableResponse<Response> response = addLineStationRequest(1L, sectionRequest, accessToken);
+
+        assertThat(response.statusCode()).isEqualTo(400);
+        assertThat((String) response.jsonPath().get("error")).isEqualTo("BOTH_STATION_ALREADY_REGISTERED_TO_LINE");
+    }
+
+    @Test
+    @DisplayName("두 역 다 노선에 존재하지 않는 경우의 구간 추가")
+    public void addSectionWhenBothStationsNotIncludeInLine() {
+        LineRequest request = new LineRequest("9호선", "황토색", 1L, 2L, 30);
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        createLineRequest(request, accessToken);
+        jdbcTemplate.update("insert into STATION (name) values ('third')");
+        jdbcTemplate.update("insert into STATION (name) values ('fourth')");
+
+        SectionRequest sectionRequest = new SectionRequest(3L, 4L, 10);
+        ExtractableResponse<Response> response = addLineStationRequest(1L, sectionRequest, accessToken);
+
+        assertThat(response.statusCode()).isEqualTo(400);
+        assertThat((String) response.jsonPath().get("error")).isEqualTo("BOTH_STATION_NOT_REGISTERED_IN_LINE");
+    }
+
+    @Test
+    @DisplayName("구간 거리가 0 일 때의 구간 추가")
+    public void addSectionWhenInvalidDistance() {
+        LineRequest request = new LineRequest("9호선", "황토색", 1L, 2L, 30);
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        createLineRequest(request, accessToken);
+
+        SectionRequest sectionRequest = new SectionRequest(1L, 2L, 0);
+        ExtractableResponse<Response> response = addLineStationRequest(1L, sectionRequest, accessToken);
+
+        assertThat(response.statusCode()).isEqualTo(400);
+        assertThat((String) response.jsonPath().get("error")).isEqualTo("INVALID_DISTANCE");
+    }
+
+    @Test
+    @DisplayName("중간 삽입인데 그 거리가 기존의 구간보다 높거나 같은 경우의 구간 삽입")
+    public void addSectionWhenInvalidDistance2() {
+        LineRequest request = new LineRequest("9호선", "황토색", 1L, 2L, 30);
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        createLineRequest(request, accessToken);
+        jdbcTemplate.update("insert into STATION (name) values ('third')");
+
+        SectionRequest sectionRequest = new SectionRequest(1L, 3L, 30);
+        ExtractableResponse<Response> response = addLineStationRequest(1L, sectionRequest, accessToken);
+
+        assertThat(response.statusCode()).isEqualTo(400);
+        assertThat((String) response.jsonPath().get("error")).isEqualTo("IMPOSSIBLE_DISTANCE");
+    }
+
+    @Test
+    @DisplayName("정상적인 구간 추가")
+    public void addSection() {
+        LineRequest request = new LineRequest("9호선", "황토색", 1L, 2L, 30);
+        String accessToken = loginRequest(new TokenRequest("email@email.com", "password")).jsonPath().get("accessToken");
+        createLineRequest(request, accessToken);
+        jdbcTemplate.update("insert into STATION (name) values ('third')");
+        SectionRequest sectionRequest = new SectionRequest(1L, 3L, 10);
+        addLineStationRequest(1L, sectionRequest, accessToken);
+
+        Integer distance = jdbcTemplate.queryForObject("select distance from SECTION where line_id = 1 and up_station_id = 1 and down_station_id =3", Integer.class);
+        assertThat(distance).isEqualTo(10);
+        Integer distance2 = jdbcTemplate.queryForObject("select distance from SECTION where line_id = 1 and up_station_id = 3 and down_station_id =2", Integer.class);
+        assertThat(distance2).isEqualTo(20);
+    }
+
+    private ExtractableResponse<Response> createLineRequest(LineRequest lineRequest, String accessToken) {
+        return RestAssured.given()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(params)
-                .when().post("/lines")
-                .then().log().all().
-                        extract();
+                .auth().oauth2(accessToken)
+                .body(lineRequest)
+                .when()
+                .post("/api/lines")
+                .then().extract();
     }
 
-    private static ExtractableResponse<Response> 지하철_노선_목록_조회_요청() {
-        return RestAssured
-                .given().log().all()
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/lines")
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> 지하철_노선_조회_요청(LineResponse response) {
-        return RestAssured
-                .given().log().all()
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/lines/{lineId}", response.getId())
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> 지하철_노선_수정_요청(LineResponse response, LineRequest params) {
-
-        return RestAssured
-                .given().log().all()
+    private ExtractableResponse<Response> findAllRequest(String accessToken) {
+        return RestAssured.given()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(params)
-                .when().put("/lines/" + response.getId())
-                .then().log().all()
-                .extract();
+                .auth().oauth2(accessToken)
+                .when()
+                .get("/api/lines")
+                .then().extract();
     }
 
-    public static ExtractableResponse<Response> 지하철_노선_제거_요청(LineResponse lineResponse) {
-        return RestAssured
-                .given().log().all()
-                .when().delete("/lines/" + lineResponse.getId())
-                .then().log().all()
-                .extract();
+    private ExtractableResponse<Response> findByIdRequest(Long id, String accessToken) {
+        return RestAssured.given()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .auth().oauth2(accessToken)
+                .when()
+                .get("/api/lines/" + id)
+                .then().extract();
     }
 
-    public static void 지하철_전체_보기_응답됨(ExtractableResponse response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    private ExtractableResponse<Response> updateLineRequest(Long id, LineRequest lineRequest, String accessToken) {
+        return RestAssured.given()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(lineRequest)
+                .auth().oauth2(accessToken)
+                .when()
+                .put("/api/lines/" + id)
+                .then().extract();
     }
 
-    public static void 지하철_전체_보기_일치함(ExtractableResponse response, List<LineMapResponse> expected) {
-        List<LineMapResponse> actual = response.jsonPath().getList(".", LineMapResponse.class);
-        assertThat(actual)
-                .usingRecursiveComparison()
-                .isEqualTo(expected);
+    private ExtractableResponse<Response> deleteLineRequest(Long id, String accessToken) {
+        return RestAssured.given()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .auth().oauth2(accessToken)
+                .when()
+                .delete("/api/lines/" + id)
+                .then().extract();
     }
 
-    public static void 지하철_노선_생성됨(ExtractableResponse response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-        assertThat(response.header("Location")).isNotBlank();
-    }
-
-    public static void 지하철_노선_생성_실패됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-    }
-
-    public static void 지하철_노선_목록_응답됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    public static void 지하철_노선_응답됨(ExtractableResponse<Response> response, LineResponse lineResponse) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-        LineResponse resultResponse = response.as(LineResponse.class);
-        assertThat(resultResponse.getId()).isEqualTo(lineResponse.getId());
-    }
-
-    public static void 지하철_노선_목록_포함됨(ExtractableResponse<Response> response, List<TotalLineResponse> createdResponses) {
-        List<Long> expectedLineIds = createdResponses.stream()
-                .map(TotalLineResponse::getId)
-                .collect(Collectors.toList());
-
-        List<Long> resultLineIds = response.jsonPath().getList(".", TotalLineResponse.class).stream()
-                .map(TotalLineResponse::getId)
-                .collect(Collectors.toList());
-
-        assertThat(resultLineIds).containsAll(expectedLineIds);
-    }
-
-    public static void 지하철_노선_수정됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    public static void 지하철_노선_삭제됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+    private ExtractableResponse<Response> addLineStationRequest(Long lineId, SectionRequest request, String accessToken) {
+        return RestAssured.given()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .auth().oauth2(accessToken)
+                .body(request)
+                .when()
+                .post("/api/lines/" + lineId + "/sections")
+                .then().extract();
     }
 }
